+++ tests/test_scenarios.py
@@ -4,8 +4,6 @@ Verify that ../examples/*/setup.py behave as expected
 
 import pytest
 
-import setupmeta
-
 from . import scenarios
 
 
@@ -17,9 +15,6 @@ def scenario_folder(request):
 
 def test_scenario(scenario_folder):
     """ Check that 'scenario' yields expected explain output """
-    if setupmeta.WINDOWS:
-        if "complex" in scenario_folder:
-            return
     scenario = scenarios.Scenario(scenario_folder)
     expected = scenario.expected_contents()
     output = scenario.replay()